Add minimap icons functionality to HBD
Since GetPlayerMapPosition returns coordinates of the currently
displayed world map, it's not possible to obtain the correct position
for minimap icons when the displayed map differs from the map
the player is on. Calling SetMapToCurrentZone and then reverting to the
previous map will trigger the WORLD_MAP_UPDATE event multiple times,
which is not a viable option in an OnUpdate loop for performance reasons.